% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_sum_scoring.R
\name{add_scores_to_sce}
\alias{add_scores_to_sce}
\title{Add Scores to SingleCellExperiment Object}
\usage{
add_scores_to_sce(sce_obj, scores, prefix = "")
}
\arguments{
\item{sce_obj}{SingleCellExperiment object}

\item{scores}{Data.frame of scores}

\item{prefix}{Prefix for colData column names (default: "")}
}
\value{
SingleCellExperiment object with scores in colData
}
\description{
Convenience function to add scores to SCE colData
}
