% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_sum_scoring.R
\name{calculate_weighted_scores}
\alias{calculate_weighted_scores}
\title{Calculate Weighted Sum Scores}
\usage{
calculate_weighted_scores(
  expression_matrix,
  reference_data,
  z_score_cutoff = 2,
  pseudobulk = FALSE,
  score_name = "weighted_sum_score",
  verbose = TRUE
)
}
\description{
Core function to calculate weighted sum of expression and z-scores
}
\keyword{internal}
