% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_expression.R
\name{get_cancer_type_from_dataset}
\alias{get_cancer_type_from_dataset}
\title{Get Cancer Type from Dataset Info}
\usage{
get_cancer_type_from_dataset(dataset, reference)
}
\description{
Get Cancer Type from Dataset Info
}
\keyword{internal}
