% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_gene_coverage}
\alias{get_gene_coverage}
\title{Get Gene Coverage}
\usage{
get_gene_coverage(genes, reference = NULL)
}
\arguments{
\item{genes}{Character vector of gene names}

\item{reference}{Reference dataset(s) to check. If NULL, checks all.}
}
\value{
Data.frame with gene coverage statistics
}
\description{
Check which genes from your data are covered in reference datasets
}
\examples{
\dontrun{
# Check gene coverage
my_genes <- c("TP53", "MYC", "EGFR", "BRCA1")
get_gene_coverage(my_genes)
}

}
