% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognostic_analysis.R
\name{get_or_create_seurat_for_markers}
\alias{get_or_create_seurat_for_markers}
\title{Get existing Seurat object or create one from matrix/SCE; add prognostic group to metadata}
\usage{
get_or_create_seurat_for_markers(expression, expr_info, group_vec, assay, slot)
}
\description{
Get existing Seurat object or create one from matrix/SCE; add prognostic group to metadata
}
\keyword{internal}
