% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_expression.R
\name{get_reference_data}
\alias{get_reference_data}
\title{Get Reference Prognostic Data}
\usage{
get_reference_data(reference, cancer_type, use_dataset_info, dataset)
}
\description{
Get Reference Prognostic Data
}
\keyword{internal}
