% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenoMap.R
\name{get_reference_genes_for_extraction}
\alias{get_reference_genes_for_extraction}
\title{Get genes to extract for memory-efficient scoring (Seurat/SCE)}
\usage{
get_reference_genes_for_extraction(reference_data, z_score_cutoff)
}
\description{
Get genes to extract for memory-efficient scoring (Seurat/SCE)
}
\keyword{internal}
