% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognostic_analysis.R
\name{infer_cell_type_column}
\alias{infer_cell_type_column}
\title{Infer which column contains cell type labels}
\usage{
infer_cell_type_column(group_labels, id_col, group_column)
}
\description{
Infer which column contains cell type labels
}
\keyword{internal}
