% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognostic_analysis.R
\name{infer_group_labels_columns}
\alias{infer_group_labels_columns}
\title{Infer helper columns from a group_labels data.frame}
\usage{
infer_group_labels_columns(group_labels, cell_id_column, group_column)
}
\description{
Infer helper columns from a group_labels data.frame
}
\keyword{internal}
