% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_cancer_types}
\alias{list_cancer_types}
\title{List Available Cancer Types}
\usage{
list_cancer_types(reference = NULL)
}
\arguments{
\item{reference}{Reference dataset name ("precog", "tcga", "pediatric_precog",
"ici_precog"). If NULL, shows all.}
}
\value{
Named list of cancer types for each reference
}
\description{
Show available cancer types for each reference dataset
}
\examples{
\dontrun{
# List all cancer types
list_cancer_types()

# List only TCGA cancer types
list_cancer_types("tcga")
}

}
