% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_sum_scoring.R
\name{normalize_scores}
\alias{normalize_scores}
\title{Calculate Z-Scores}
\usage{
normalize_scores(scores)
}
\arguments{
\item{scores}{Numeric vector of scores}
}
\value{
Numeric vector of z-scores
}
\description{
Normalize scores to z-scores
}
