% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognostic_data.R
\docType{data}
\name{pediatric}
\alias{pediatric}
\title{Pediatric Cancer Prognostic Z-Scores}
\format{
A data.frame with genes as rows and pediatric cancer types as columns.
}
\usage{
pediatric
}
\description{
Gene-level prognostic z-scores for pediatric cancers.
}
\keyword{datasets}
