% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_score_distribution}
\alias{plot_score_distribution}
\title{Plot Score Distribution}
\usage{
plot_score_distribution(
  scores,
  score_column = NULL,
  main = "Score Distribution"
)
}
\arguments{
\item{scores}{Numeric vector or data.frame of scores}

\item{score_column}{If scores is data.frame, which column to plot}

\item{main}{Plot title}
}
\description{
Simple histogram of score distribution
}
