% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognostic_analysis.R
\name{process_expression_for_markers}
\alias{process_expression_for_markers}
\title{Process expression input for marker finding}
\usage{
process_expression_for_markers(expression, assay = NULL, slot = "data")
}
\description{
Process expression input for marker finding
}
\keyword{internal}
