% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handlers.R
\name{process_expression_input}
\alias{process_expression_input}
\title{Process Expression Input}
\usage{
process_expression_input(
  expression,
  pseudobulk = FALSE,
  group_by = NULL,
  assay = NULL,
  slot = "data",
  genes_to_extract = NULL,
  verbose = TRUE
)
}
\description{
Convert various expression data formats to matrix
}
\keyword{internal}
