% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handlers.R
\name{process_sce}
\alias{process_sce}
\title{Process SingleCellExperiment Object}
\usage{
process_sce(obj, pseudobulk, group_by, assay)
}
\description{
Process SingleCellExperiment Object
}
\keyword{internal}
