% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handlers.R
\name{process_spatial_experiment}
\alias{process_spatial_experiment}
\title{Process SpatialExperiment Object}
\usage{
process_spatial_experiment(obj, pseudobulk, group_by, assay)
}
\description{
Process SpatialExperiment Object
}
\keyword{internal}
