% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognostic_analysis.R
\name{resolve_labels_from_df}
\alias{resolve_labels_from_df}
\title{Resolve a label column from group_labels data.frame to match cell_ids}
\usage{
resolve_labels_from_df(group_labels, cell_ids, id_col, label_column)
}
\description{
Resolve a label column from group_labels data.frame to match cell_ids
}
\keyword{internal}
