% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognostic_analysis.R
\name{run_markers_on_matrix_by_celltype}
\alias{run_markers_on_matrix_by_celltype}
\title{Run marker detection on a matrix by cell type}
\usage{
run_markers_on_matrix_by_celltype(
  mat,
  group_vec,
  cell_type_vec,
  min.pct = 0.1,
  logfc.threshold = 0.25,
  pval_threshold = 0.05,
  max_cells_per_ident = 5000L,
  verbose = TRUE
)
}
\description{
For each cell type, compute markers for:
\itemize{
\item Most Adverse within that cell type vs all other groups within that cell type
\item Most Favorable within that cell type vs all other groups within that cell type
}
}
\keyword{internal}
