% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognostic_analysis.R
\name{standardize_findmarkers_output}
\alias{standardize_findmarkers_output}
\title{Convert FindMarkers result to standard columns and filter by p-value}
\usage{
standardize_findmarkers_output(df, pval_threshold)
}
\description{
Convert FindMarkers result to standard columns and filter by p-value
}
\keyword{internal}
