% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handlers.R
\name{validate_expression_matrix}
\alias{validate_expression_matrix}
\title{Validate Expression Matrix}
\usage{
validate_expression_matrix(mat)
}
\description{
Validate Expression Matrix
}
\keyword{internal}
