% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_package_data}
\alias{validate_package_data}
\title{Validate Package Data}
\usage{
validate_package_data()
}
\description{
Check that all required data objects are present and valid
}
\keyword{internal}
