% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_reference.R
\name{z_scores_binary}
\alias{z_scores_binary}
\title{Z-scores from logistic regression per gene (binary outcome)}
\usage{
z_scores_binary(
  expr,
  group,
  verbose = TRUE,
  binary_positive_reference = c("second", "first")
)
}
\description{
Z-scores from logistic regression per gene (binary outcome)
}
\keyword{internal}
