% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_sum_scoring.R
\name{add_scores_to_seurat}
\alias{add_scores_to_seurat}
\title{Add Scores to Seurat Object}
\usage{
add_scores_to_seurat(seurat_obj, scores, prefix = "")
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{scores}{Data.frame of scores (output from score_expression)}

\item{prefix}{Prefix for metadata column names (default: "")}
}
\value{
Seurat object with scores added to metadata
}
\description{
Convenience function to add scores directly to Seurat metadata
}
