% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_sum_scoring.R
\name{compute_scores}
\alias{compute_scores}
\title{Compute Score Vectors}
\usage{
compute_scores(
  expression_data,
  prognostic_scores,
  pseudobulk = FALSE,
  verbose = TRUE
)
}
\description{
Efficiently compute weighted sum using matrix operations
}
\keyword{internal}
